Added EventMachine Sinatra Example
When I first implemented this, it took me a while to figure out
how to do it properly. Even though it rides the line between being
an EventMachine example and a Sinatra example, I believe it can
be very useful to other Sinatra users. And more examples can't hurt
right?